+++ CHANGES.md
@@ -2,6 +2,7 @@
 
   * Support for non-expiring scope
   * Cancellation of uploads
+  * Added Request#withFile(String, byte[]), Request#withFile(String, ByteBuffer)
 
 ## 1.0.0 2011-05-19